+++ packages/constructs/L3/analytics/quicksight-account-l3-construct/test/quicksight-account.compliance.test.ts
@@ -10,7 +10,7 @@ import { Match, Template } from "aws-cdk-lib/assertions";
 import { Protocol } from "aws-cdk-lib/aws-ec2";
 import { QuickSightAccountL3Construct, QuickSightAccountL3ConstructProps } from "../lib";
 
-describe( 'CAEF Compliance Stack Tests', () => {
+describe( 'QS Account Mandatory Tests', () => {
   const testApp = new CaefTestApp()
   const testQSSecurityGroup: CaefSecurityGroupRuleProps = {
     sg: [
@@ -18,9 +18,19 @@ describe( 'CAEF Compliance Stack Tests', () => {
         sgId: "sg-1234abcd",
         port: 1111,
         protocol: Protocol.TCP
-      },
+      }
+    ],
+    ipv4: [
       {
-        sgId: "sg-1234abcd",
+        cidr: "10.0.0.0/32",
+        port: 1000,
+        toPort: 2000,
+        protocol: Protocol.TCP
+      }
+    ],
+    prefixList: [
+      {
+        prefixList: "pl-abc123",
         port: 1000,
         toPort: 2000,
         protocol: Protocol.TCP
@@ -39,7 +49,14 @@ describe( 'CAEF Compliance Stack Tests', () => {
       emailAddress: "test@example.com",
       contactNumber: "1234546879",
       vpcId: "vpc-abcd1234",
-      glueResourceAccess: [ "database/some-database-name*" ]
+      subnetIds: [ "test-subnet-id1", "test-subnet-id2" ],
+      glueResourceAccess: [ "database/some-database-name*" ],
+      ipRestrictions: [ {
+        cidr: "1.1.1.1/1",
+        description: "testing1"
+      }, {
+        cidr: "2.2.2.2/2"
+      } ]
     },
     naming: testApp.naming,
 
@@ -49,10 +66,8 @@ describe( 'CAEF Compliance Stack Tests', () => {
   new QuickSightAccountL3Construct( testApp.testStack, 'test-stack', constructProps )
   testApp.checkCdkNagCompliance( testApp.testStack )
   const template = Template.fromStack( testApp.testStack )
-  test( 'Validate if the 2 Lambda functions[CR and QSAccount] are created', () => {
-    template.resourceCountIs( "AWS::Lambda::Function", 2 );
-  } );
-  // Verify Properties for QS Folders like folderpermissions, folderName, etc.
+  // console.log( JSON.stringify( template, undefined, 2 ) )
+
   test( 'Test QS Account With Sample Config', () => {
     template.hasResourceProperties( "AWS::CloudFormation::CustomResource", {
       "accountDetail": Match.objectLike( {
@@ -61,6 +76,102 @@ describe( 'CAEF Compliance Stack Tests', () => {
         notificationEmail: "test@example.com",
         accountName: "test-org-test-env-test-domain-test-module"
       } ),
+    } )
+  } );
+
+
+  test( 'Test IP Restrictions', () => {
+    template.hasResourceProperties( "Custom::ip-restrictions", {
+      "accountId": "test-account",
+      "ipRestrictionsMap": {
+        "1.1.1.1/1": "testing1",
+        "2.2.2.2/2": "Restriction for 2.2.2.2/2"
+      }
+    } )
+  } );
+
+
+  test( 'Test Security Group Ingress', () => {
+    template.hasResourceProperties( "AWS::EC2::SecurityGroupIngress", {
+      "IpProtocol": "tcp",
+      "CidrIp": "10.0.0.0/32",
+      "Description": "from 10.0.0.0/32:tcp RANGE 1-65535",
+      "FromPort": 1,
+      "GroupId": {
+        "Fn::GetAtt": [
+          "teststackquicksightsgB57FA8A2",
+          "GroupId"
+        ]
+      },
+      "ToPort": 65535
+    } )
+    template.hasResourceProperties( "AWS::EC2::SecurityGroupIngress", {
+      "IpProtocol": "tcp",
+      "Description": "from sg-1234abcd:tcp RANGE 1-65535",
+      "FromPort": 1,
+      "GroupId": {
+        "Fn::GetAtt": [
+          "teststackquicksightsgB57FA8A2",
+          "GroupId"
+        ]
+      },
+      "SourceSecurityGroupId": "sg-1234abcd",
+      "ToPort": 65535
+    } )
+    template.hasResourceProperties( "AWS::EC2::SecurityGroupIngress", {
+      "IpProtocol": "tcp",
+      "Description": "from pl-abc123:tcp RANGE 1-65535",
+      "FromPort": 1,
+      "GroupId": {
+        "Fn::GetAtt": [
+          "teststackquicksightsgB57FA8A2",
+          "GroupId"
+        ]
+      },
+      "SourcePrefixListId": "pl-abc123",
+      "ToPort": 65535
+    } )
+  } );
+
+  test( 'Test Security Group Egress', () => {
+    template.hasResourceProperties( "AWS::EC2::SecurityGroupEgress", {
+      "GroupId": {
+        "Fn::GetAtt": [
+          "teststackquicksightsgB57FA8A2",
+          "GroupId"
+        ]
+      },
+      "IpProtocol": "tcp",
+      "CidrIp": "10.0.0.0/32",
+      "Description": "to 10.0.0.0/32:tcp RANGE 1000-2000",
+      "FromPort": 1000,
+      "ToPort": 2000
+    } )
+    template.hasResourceProperties( "AWS::EC2::SecurityGroupEgress", {
+      "GroupId": {
+        "Fn::GetAtt": [
+          "teststackquicksightsgB57FA8A2",
+          "GroupId"
+        ]
+      },
+      "IpProtocol": "tcp",
+      "Description": "to sg-1234abcd:tcp PORT 1111",
+      "DestinationSecurityGroupId": "sg-1234abcd",
+      "FromPort": 1111,
+      "ToPort": 1111
+    } )
+    template.hasResourceProperties( "AWS::EC2::SecurityGroupEgress", {
+      "GroupId": {
+        "Fn::GetAtt": [
+          "teststackquicksightsgB57FA8A2",
+          "GroupId"
+        ]
+      },
+      "IpProtocol": "tcp",
+      "Description": "to pl-abc123:tcp RANGE 1000-2000",
+      "DestinationPrefixListId": "pl-abc123",
+      "FromPort": 1000,
+      "ToPort": 2000
     } )
   } );
 } );